AG-12895 Join SeriesAreaManager and SeriesAreaClickManager
